hfa-core/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hfa-core"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7",
    "cryptography>=42",
    "opentelemetry-api>=1.24",
]

[project.optional-dependencies]
nats  = ["nats-py>=2.6"]
kafka = ["aiokafka>=0.10"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
filterwarnings = ["ignore::DeprecationWarning", "ignore::pytest.PytestUnraisableExceptionWarning"]
